+++ arch/riscv/cpu/generic/include/arch_io.h
@@ -246,9 +246,9 @@ static inline void arch_outsl(unsigned long p, const void *b, int c)
 #define arch_out_le32(a, v)	{__io_bw(); __raw_write32(a, v); __io_aw(); }
 #define arch_in_be32(a)		({u32 v; __io_br(); v = __raw_read32(a); __io_ar(); rev32(v); })
 #define arch_out_be32(a, v)	{__io_bw(); __raw_write32(a, rev32(v)); }
-#define arch_in_le64(a)		({u32 v; __io_br(); v = __raw_read64(a); __io_ar(); v; })
+#define arch_in_le64(a)		({u64 v; __io_br(); v = __raw_read64(a); __io_ar(); v; })
 #define arch_out_le64(a, v)	{__io_bw(); __raw_write64(a, v); __io_aw(); }
-#define arch_in_be64(a)		({u32 v; __io_br(); v = __raw_read64(a); __io_ar(); rev64(v); })
+#define arch_in_be64(a)		({u64 v; __io_br(); v = __raw_read64(a); __io_ar(); rev64(v); })
 #define arch_out_be64(a, v)	{__io_bw(); __raw_write64(a, rev64(v)); __io_aw(); }
 
 #define arch_in_8_relax(a)		({u8 v; __io_rbr(); v = __raw_read8(a); __io_rar(); v; })
@@ -261,9 +261,9 @@ static inline void arch_outsl(unsigned long p, const void *b, int c)
 #define arch_out_le32_relax(a, v)	{__io_rbw(); __raw_write32(a, v); __io_raw(); }
 #define arch_in_be32_relax(a)		({u32 v; __io_rbr(); v = __raw_read32(a); __io_rar(); rev32(v); })
 #define arch_out_be32_relax(a, v)	{__io_rbw(); __raw_write32(a, rev32(v)); __io_raw(); }
-#define arch_in_le64_relax(a)		({u32 v; __io_rbr(); v = __raw_read64(a); __io_rar(); v; })
+#define arch_in_le64_relax(a)		({u64 v; __io_rbr(); v = __raw_read64(a); __io_rar(); v; })
 #define arch_out_le64_relax(a, v)	{__io_rbw(); __raw_write64(a, v); __io_raw(); }
-#define arch_in_be64_relax(a)		({u32 v; __io_rbr(); v = __raw_read64(a); __io_rar(); rev64(v); })
+#define arch_in_be64_relax(a)		({u64 v; __io_rbr(); v = __raw_read64(a); __io_rar(); rev64(v); })
 #define arch_out_be64_relax(a, v)	{__io_rbw(); __raw_write64(a, rev64(v)); __io_raw(); }
 
 #endif
